refactor(agent): split policy engine responsibilities
